utils: move the guess routines to sherpa.utils.guess
This doesn't really improve anything, but does break up the
utils module more (following recent changes for random and
parallel).

The routines are still available from sherpa.utils but that will
hopefully change in the future. One issue is how to warn the user
that they are using an "outdated" symbol.
