+++ docs/overview/guess.rst
@@ -1,0 +1,35 @@
+*****************************
+The sherpa.utils.guess module
+*****************************
+
+.. currentmodule:: sherpa.utils.guess
+
+.. automodule:: sherpa.utils.guess
+
+   .. rubric:: Constants
+
+   .. autosummary::
+      :toctree: api
+
+      _guess_ampl_scale
+
+   .. rubric:: Functions
+
+   .. autosummary::
+      :toctree: api
+
+      get_midpoint
+      get_peak
+      get_valley
+      get_fwhm
+      guess_fwhm
+      param_apply_limits
+      get_amplitude_position
+      guess_amplitude
+      guess_amplitude_at_ref
+      guess_amplitude2d
+      guess_reference
+      get_position
+      guess_position
+      guess_bounds
+      guess_reference
